订单ID和用户ID复合分片算法的实现

+++ src/main/java/com/sharding/mysql/example/entity/Order.java
@@ -20,13 +20,13 @@ public class Order {
      */
     @Id
     @Column(name = "order_id")
-    private Integer orderId;
+    private Long orderId;
 
     /**
      * 用户流水ID
      */
     @Column(name = "user_Id")
-    private Integer userId;
+    private Long userId;
 
     /**
      * 订单编号
@@ -55,19 +55,19 @@ public class Order {
         this.orderItemList = orderItemList;
     }
 
-    public Integer getOrderId() {
+    public Long getOrderId() {
         return orderId;
     }
 
-    public void setOrderId(Integer orderId) {
+    public void setOrderId(Long orderId) {
         this.orderId = orderId;
     }
 
-    public Integer getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
-    public void setUserId(Integer userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
